support "reference" attribute in tag.consumer
Consumer tags now allow an optional attribute "reference".
By default this attribute is `true` and keeps compatibility with previous definitions.
When this is `false`, service ids will be injected to the consumer, not service references.
This may help in case you're making a service locator, for example.

Kudos to @andresroro for the idea